[IR Runtime] add more path_expand uts
